+++ src/gracie.zig
@@ -8,37 +8,40 @@ const c = @cImport({
 
 const debug = std.debug;
 const allocator = std.mem.Allocator;
+const fixed_buffer_allocator = std.heap.FixedBufferAllocator;
 const array_list = std.ArrayList;
 
 const self = @This();
 
 // NOTE(cjb): Keep in mind this will need to be exposed at some point or another.
-const gracie_match = struct
+const match = struct
 {
     SO: c_ulonglong,
     EO: c_ulonglong,
-    CatID: c_uint,
+    ID: c_uint,
 };
 
 const loaded_py_module = struct
 {
-//    ModName: []u8,
-    CatName: []u8,
-    CatID: c_uint,
+    NameZ: []u8,
 };
 
 // TODO(cjb): Loaded extractors buf...
+const cat_box = struct
+{
+    CategoryName: []u8,
+    MainPyModuleIndex: usize,
+    StartPatternID: c_uint,
+    EndPatternID: c_uint, // Exclusive
+};
 
 Database: ?*c.hs_database_t,
+nDatabaseBytes: usize,
 Scratch: ?*c.hs_scratch_t,
 Ally: allocator,
+FBackBuf: []u8,
 
-//ManagedVT: gracie_managed_vtable,
-
-/// List of category ids, where pattern id is the index of the associated category id.
-CatIDs: array_list(c_uint),
-MatchList: array_list(gracie_match), // TODO(cjb): Benchmark init. of this data strcuture within
-                                     // extract call instead of GracieInit
+CatBoxes: array_list(cat_box),
 LoadedPyModules: array_list(loaded_py_module),
 //
 // Possible status codes that may be returned during gracie calls.
@@ -109,92 +112,10 @@ fn ErrToCode(Err: anyerror) c_int
     }
 }
 
-//const gracie_free_fn = *const fn(Ptr: ?*anyopaque) callconv(.C) void;
-//const gracie_resize_fn = *const fn(Ptr: ?*anyopaque, Size: usize) callconv(.C) ?*anyopaque;
-//
-//const gracie_managed_vtable = extern struct
-//{
-//    Alloc: gracie_alloc_fn,
-//    Free: gracie_free_fn,
-//    Resize: gracie_resize_fn,
-//};
-
-//
-// Allocator compatable vtable functions for calling c memory procedures *mostly* ripped straight
-// from the zig std library.
-//
-
-//fn ManagedAlloc(Ctx: *anyopaque, Len: usize, Log2PtrAlign: u8, RetAddr: usize) ?[*]u8
-//{
-//    _ = RetAddr;
-//    debug.assert(Log2PtrAlign <= comptime std.math.log2_int(usize, @alignOf(std.c.max_align_t)));
-//    const VT = @ptrCast(*gracie_managed_vtable, @alignCast(@alignOf(gracie_managed_vtable), Ctx));
-//    return @ptrCast(?[*]u8, VT.Alloc(Len));
-//}
-//
-//fn GetRecordPtr(Buf: []u8) *align(1) usize
-//{
-//    return @intToPtr(*align(1) usize, @ptrToInt(Buf.ptr) + Buf.len);
-//}
-//
-//fn ManagedResize(Ctx: *anyopaque, Buf: []u8, Log2OldAlign: u8, NewLen: usize, RetAddr: usize) ?[*]u8
-//{
-//    _ = Log2OldAlign;
-//    _ = RetAddr;
-//    const VT = @ptrCast(*gracie_managed_vtable, @alignCast(@alignOf(gracie_managed_vtable), Ctx));
-//
-//    return @ptrCast(?[*]u8, VT.Resize(Buf.ptr, NewLen));
-//    assert(new_ptr == @intToPtr(*anyopaque, root_addr));
-//    getRecordPtr(buf.ptr[0..new_size]).* = root_addr;
-//}
-//
-//fn ManagedFree(Ctx: *anyopaque, Buf: []u8, Log2OldAlign: u8, NewLen: usize, RetAddr: usize) void
-//{
-//    _ = Log2OldAlign;
-//    _ = RetAddr;
-//    _ = NewLen;
-//    const VT = @ptrCast(*gracie_managed_vtable, @alignCast(@alignOf(gracie_managed_vtable), Ctx));
-//    VT.Free(Buf.ptr);
-//}
-//export fn GracieManagedInit(Ctx: ?*?*anyopaque, Alloc: gracie_alloc_fn, Free: gracie_free_fn,
-//    Resize: gracie_resize_fn, ArtifactPathZ: ?[*:0]const u8) callconv(.C) c_int
-//{
-//    var Self = @ptrCast(?*?*self, @alignCast(@alignOf(?*self), Ctx));
-//
-//    // Stack allocator from stack vtable
-//    var SVT = gracie_managed_vtable{
-//        .Alloc = Alloc,
-//        .Resize= Resize,
-//        .Free = Free,
-//    };
-//    var SAlly = allocator{
-//        .ptr = @ptrCast(*anyopaque, &SVT),
-//        .vtable = &.{
-//            .alloc = ManagedAlloc,
-//            .resize = ManagedResize,
-//            .free = ManagedFree,
-//        },
-//    };
-//
-//    // Perform managed alloc storing result of self.Init in allocated slot.
-//    var SelfBytes = ManagedAlloc(&SVT, @sizeOf(self), @sizeOf(*void * 2), 0);
-//    Self.?.* = @ptrCast(?*self, @alignCast(@alignOf(self), SelfBytes));
-//    Self.?.* orelse return GRACIE_NOMEM; // Check to see if actually alloc'd.
-//    Self.?.*.?.* = Init(Ctx, SAlly, ArtifactPathZ) catch |Err|
-//        return ErrToCode(Err);
-//
-//    // Copy stack allocated vtable to new Self's vtable then update stack allocated allocator's
-//    //  pointer to point to Self's vtable (just copied) instead of the original.
-//    Self.ManagedVT = SVT;
-//    Self.Ally.ptr = &Self.ManagedVT;
-//
-//    return GRACIE_SUCCESS;
-//}
-
 export fn GracieInit(Ctx: ?*?*anyopaque, ArtifactPathZ: ?[*:0]const u8) callconv(.C) c_int
 {
     var Self = @ptrCast(?*?*self, @alignCast(@alignOf(?*self), Ctx));
-    var Ally = std.heap.c_allocator;
+    var Ally = std.heap.page_allocator;
     Self.?.* = Ally.create(self) catch |Err|
         return ErrToCode(Err);
     Self.?.*.?.* = Init(Ally, ArtifactPathZ) catch |Err|
@@ -205,15 +126,7 @@ export fn GracieInit(Ctx: ?*?*anyopaque, ArtifactPathZ: ?[*:0]const u8) callconv
 pub fn Init(Ally: allocator, ArtifactPathZ: ?[*:0]const u8) !self
 {
     var Self: self = undefined;
-
-    // Copy alloactor
     Self.Ally = Ally;
-
-    // Initialize sempy ( this must be initialized before any sempy fns are called. )
-    try sempy.Init();
-
-    // Initialize list to keep track of loaded sempy modules.
-    Self.LoadedPyModules = array_list(loaded_py_module).init(Self.Ally);
 
 //
 // Deserialize artifact
@@ -224,14 +137,52 @@ pub fn Init(Ally: allocator, ArtifactPathZ: ?[*:0]const u8) !self
     const ArtiF = try std.fs.cwd().openFile(ArtifactPathZ.?[0..PathLen], .{});
 
     // Read artifact header
-    const ArtiHeader = try ArtiF.reader().readStruct(common.gracie_arti_header);
+    const ArtiHeader = try ArtiF.reader().readStruct(common.arti_header);
+
+//
+// Read python module data and pass it to sempy for loading.
+//
+    // Sempy must be initialized before any sempy fns are called.
+    try sempy.Init();
+
+    // Future sempy run calls will require module name so we will store them here.
+    Self.LoadedPyModules = array_list(loaded_py_module).init(Self.Ally);
+
+    var PyModIndex: usize = 0;
+    while (PyModIndex < ArtiHeader.nPyModules) : (PyModIndex += 1)
+    {
+        const PyModHeader = try ArtiF.reader().readStruct(common.arti_py_module_header);
+
+        // Make loaded module & new NameZ buf.
+        var LoadedMod = loaded_py_module{
+            .NameZ = try Self.Ally.alloc(u8, PyModHeader.nPyNameBytes + 1),
+        };
+
+        // Code buf
+        var SourceZ = try Self.Ally.alloc(u8, PyModHeader.nPySourceBytes + 1);
+        defer Self.Ally.free(SourceZ); // Uneeded after sempy load call.
+
+        // Read name & code
+        debug.assert(try ArtiF.readAll(LoadedMod.NameZ[0 .. PyModHeader.nPyNameBytes]) ==
+            PyModHeader.nPyNameBytes);
+        debug.assert(try ArtiF.readAll(SourceZ[0 .. PyModHeader.nPySourceBytes]) ==
+            PyModHeader.nPySourceBytes);
+
+        // Null terminate both name & code buffers.
+        LoadedMod.NameZ[LoadedMod.NameZ.len - 1] = 0;
+        SourceZ[SourceZ.len - 1] = 0;
+
+        // Attempt to load module & append to module list.
+        try sempy.LoadModuleFromSource(LoadedMod.NameZ, SourceZ);
+        try Self.LoadedPyModules.append(LoadedMod);
+    }
 
     // Read n extractor definitions
     var ExtractorDefIndex: usize = 0;
     while (ExtractorDefIndex < ArtiHeader.nExtractorDefs) : (ExtractorDefIndex += 1)
     {
         // Read definition header
-        const DefHeader = try ArtiF.reader().readStruct(common.gracie_extractor_def_header);
+        const DefHeader = try ArtiF.reader().readStruct(common.arti_def_header);
 
         // TODO(cjb): Store these values somewhere
         var Country: [2]u8 = undefined;
@@ -241,7 +192,6 @@ pub fn Init(Ally: allocator, ArtifactPathZ: ?[*:0]const u8) !self
         // Read serialized database
         var SerializedBytes = try Self.Ally.alloc(u8, DefHeader.DatabaseSize);
         defer Self.Ally.free(SerializedBytes);
-
         debug.assert(try ArtiF.reader().readAll(SerializedBytes) == DefHeader.DatabaseSize);
 
         // TODO(cjb): hs_set_misc_allocator():
@@ -254,7 +204,7 @@ pub fn Init(Ally: allocator, ArtifactPathZ: ?[*:0]const u8) !self
         var DatabaseBuf = try Self.Ally.alloc(u8, DefHeader.DatabaseSize);
         Self.Database = @ptrCast(*c.hs_database_t, @alignCast(@alignOf(c.hs_database_t),
                 DatabaseBuf.ptr));
-
+        Self.nDatabaseBytes = DefHeader.DatabaseSize;
         try HSCodeToErr(c.hs_deserialize_database_at(SerializedBytes.ptr, SerializedBytes.len,
             Self.Database));
 
@@ -263,50 +213,42 @@ pub fn Init(Ally: allocator, ArtifactPathZ: ?[*:0]const u8) !self
         try HSCodeToErr(c.hs_alloc_scratch(Self.Database, &Self.Scratch));
 
 //
-// Initialize category id's & load py modules
+// Read category data
 //
-        Self.CatIDs = array_list(c_uint).init(Self.Ally);
+        Self.CatBoxes = array_list(cat_box).init(Ally);
+
+        var PatternSum: usize = 0;
         var CatIndex: usize = 0;
         while (CatIndex < DefHeader.nCategories) : (CatIndex += 1)
         {
-            // Read category header
-            const CatHeader = try ArtiF.reader().readStruct(common.gracie_extractor_cat_header);
+            const CatHeader = try ArtiF.reader().readStruct(common.arti_cat_header);
 
-            // Record new module
-            var LMod = loaded_py_module{
-                //.ModName= TODO(cjb): Track module name
-                .CatName = try Self.Ally.alloc(u8, CatHeader.nCategoryNameBytes + 1),
-                .CatID = @intCast(c_uint, CatIndex),
-            };
-            debug.assert(try ArtiF.readAll(LMod.CatName[0 .. CatHeader.nCategoryNameBytes]) ==
-                CatHeader.nCategoryNameBytes);
-            LMod.CatName[CatHeader.nCategoryNameBytes] = 0;
-            try Self.LoadedPyModules.append(LMod);
+            // Category name
+            var CategoryName = try Ally.alloc(u8, CatHeader.nCategoryNameBytes);
+            debug.assert(try ArtiF.readAll(CategoryName) == CatHeader.nCategoryNameBytes);
 
-            var PatternIndex: usize = 0;
-            while (PatternIndex < CatHeader.nPatterns) : (PatternIndex += 1)
-            {
-                try Self.CatIDs.append(@intCast(c_uint, CatIndex));
-            }
+            // Pattern count
+            var nPatternsForCategory: usize = undefined;
+            debug.assert(try ArtiF.readAll(@ptrCast([*]u8, &nPatternsForCategory)
+                    [0 .. @sizeOf(usize)]) == @sizeOf(usize));
+            PatternSum += nPatternsForCategory;
 
-            // Read module's source code
-            var ModSourceCode = try Self.Ally.alloc(u8, CatHeader.nPyPluginSourceBytes + 1);
-            defer Self.Ally.free(ModSourceCode); // We don't need to keep this around.
-            debug.assert(try ArtiF.readAll(ModSourceCode[0 .. CatHeader.nPyPluginSourceBytes]) ==
-                CatHeader.nPyPluginSourceBytes);
-            ModSourceCode[CatHeader.nPyPluginSourceBytes] = 0;
+            // Mainpy module index
+            var MainPyModuleIndex: usize = undefined;
+            debug.assert(try ArtiF.readAll(@ptrCast([*]u8, &MainPyModuleIndex)
+                    [0 .. @sizeOf(usize)]) == @sizeOf(usize));
 
-            // Load sempy module
-            var SMCtx = sempy.module_ctx{
-                .Source=ModSourceCode[0..ModSourceCode.len - 1 :0],
-                .Name=LMod.CatName[0..LMod.CatName.len - 1 :0],
-            };
-            try sempy.LoadModuleFromSource(&SMCtx);
+           try Self.CatBoxes.append(cat_box{
+               .CategoryName = CategoryName,
+               .MainPyModuleIndex = MainPyModuleIndex,
+               .StartPatternID = @intCast(c_uint, PatternSum - nPatternsForCategory),
+               .EndPatternID = @intCast(c_uint, PatternSum),
+           });
         }
     }
 
-    // Initialize match list
-    Self.MatchList = array_list(gracie_match).init(Self.Ally);
+    // Initialize extractor scratch buffer
+    Self.FBackBuf = try Ally.alloc(u8, 1024*20);
 
     return Self;
 }
@@ -314,36 +256,63 @@ pub fn Init(Ally: allocator, ArtifactPathZ: ?[*:0]const u8) !self
 fn EventHandler(ID: c_uint, From: c_ulonglong, To: c_ulonglong, _: c_uint,
     Ctx: ?*anyopaque) callconv(.C) c_int
 {
-    const This = @ptrCast(?*self, @alignCast(@alignOf(?*self), Ctx)) orelse unreachable;
-    if (This.MatchList.items.len + 1 > This.MatchList.capacity)
+    const MatchList = @ptrCast(?*array_list(match),
+        @alignCast(@alignOf(?*array_list(match)), Ctx)) orelse unreachable;
+    if (MatchList.items.len + 1 > MatchList.capacity)
     {
-        // NOTE(cjb): Should an error be thrown here?
-        This.MatchList.append(.{.SO=From, .EO=To, .CatID=This.CatIDs.items[ID]}) catch unreachable;
+        // TODO(cjb): Decide if this should be handled or not.
+        MatchList.append(.{.SO=From, .EO=To, .ID=ID}) catch unreachable;
         return 0;
+    }
+    else
+    {
+        unreachable; //TODO(cjb): Grow matchlist...
     }
     return 1;
 }
 
-export fn GracieExtract(Ctx: ?*?*anyopaque, Text: ?[*]u8, nTextBytes: c_uint) callconv(.C) c_int
+export fn GracieExtract(Ctx: ?*?*anyopaque, Text: ?[*]const u8,
+    nTextBytes: c_uint, Result: ?*[*]c_uint, n32sCopied: ?*c_uint) callconv(.C) c_int
 {
     var Self = @ptrCast(?*?*self, @alignCast(@alignOf(?*self), Ctx));
-    Extract(Self.?.*.?, Text, nTextBytes) catch |Err|
+    var ExtractResult = Extract(Self.?.*.?, Text.?[0 .. nTextBytes]) catch |Err|
         return ErrToCode(Err);
+
+    n32sCopied.?.* = @intCast(c_uint, ExtractResult.len);
+    Result.?.* = @ptrCast([*]c_uint, ExtractResult.ptr);
     return GRACIE_SUCCESS;
 }
 
-pub fn Extract(Self: *self, Text: ?[*]u8, nTextBytes: c_uint) !void
+pub fn Extract(Self: *self, Text: []const u8) ![]u32
 {
-    // Reset stuff
-    Self.MatchList.clearRetainingCapacity();
+    // Initialize fixed buffer allocator from scratch space.
+    var FBufAlly = fixed_buffer_allocator.init(Self.FBackBuf);
+
+    // Init MatchList
+    var MatchList = array_list(match).init(FBufAlly.allocator());
+
+    var Result: []u32 = try FBufAlly.allocator().alloc(u32, 1024*2); // 2kib for result
+    var n32sCopied: usize = 0;
 
     // Hyperscannnnn!!
-    try HSCodeToErr(c.hs_scan(Self.Database, Text, nTextBytes, 0, Self.Scratch,
-            EventHandler, Self));
-    for (Self.MatchList.items) |M|
+    try HSCodeToErr(c.hs_scan(Self.Database, Text.ptr, @intCast(c_uint, Text.len), 0, Self.Scratch,
+            EventHandler, &MatchList));
+    for (MatchList.items) |M|
     {
-        try sempy.RunModule(Text.?[M.SO .. M.EO], M.CatID);
+        var MainModuleIndex: usize = undefined;
+        for (Self.CatBoxes.items) |Cat|
+        {
+            if ((M.ID >= Cat.StartPatternID) and
+                (M.ID < Cat.EndPatternID))
+            {
+                MainModuleIndex = Cat.MainPyModuleIndex;
+            }
+        }
+        n32sCopied = try sempy.RunModule(Text[M.SO .. M.EO],
+            Self.LoadedPyModules.items[MainModuleIndex].NameZ, Result);
+        break;
     }
+    return Result[0 .. n32sCopied];
 }
 
 export fn GracieDeinit(Ctx: ?*?*anyopaque) callconv(.C) c_int
@@ -359,18 +328,38 @@ pub fn Deinit(Self: *self) !void
 {
     // Free hs scratch
     try HSCodeToErr(c.hs_free_scratch(Self.Scratch));
+    Self.Ally.free(@ptrCast([*]u8, Self.Database.?)[0 .. Self.nDatabaseBytes]);
+
+    Self.Ally.free(Self.FBackBuf);
+
+    for (Self.CatBoxes.items) |Cat| Self.Ally.free(Cat.CategoryName);
+    Self.CatBoxes.deinit();
 
     // Deinitialize sempy
-    try sempy.Deinit();
+    //try sempy.Deinit(); FIXME(cjb): whyyyyy!!!
+
+    for (Self.LoadedPyModules.items) |Mod| Self.Ally.free(Mod.NameZ);
+    Self.LoadedPyModules.deinit();
 }
 
-test "Gracie"
+test "Gracie C"
 {
-    var GracieCtx: ?*anyopaque = null;
-    try std.testing.expect(
-        self.Init(&GracieCtx, "./data/gracie.bin") ==
-        GRACIE_SUCCESS);
-    try std.testing.expect(
-        self.Deinit(&GracieCtx) ==
-        GRACIE_SUCCESS);
+    var G: ?*anyopaque = undefined;
+    debug.assert(GracieInit(&G, "./data/gracie.bin") == GRACIE_SUCCESS);
+    const Text = "Earn $30 an hour";
+    debug.assert(GracieExtract(&G, Text, Text.len) == GRACIE_SUCCESS);
+    debug.assert(GracieDeinit(&G) == GRACIE_SUCCESS);
 }
+
+//test "Gracie"
+//{
+//    var Ally = std.testing.allocator;
+//    var BackingBuf = try Ally.alloc(u8, 0x1000*9);
+//    defer Ally.free(BackingBuf);
+//    var SAlly = slab_allocator.Init(BackingBuf);
+//    var G = try self.Init(SAlly.Allocator(), "./data/gracie.bin");
+//    try self.Extract(&G, "Earn $30 an hour");
+//
+//    slab_allocator.DEBUGSlabVisularizer(&SAlly);
+//    try self.Deinit(&G);
+//}